Moonray 4848 v3 rebase (#819)
* Refactor Shader logging to avoid race conditions

Shader.cc and LoggingAssert.h are added to scene/rdl2/SConscript and CMakeLists.txt to include it in the compilation process. The main change made was in the Shader.h and Shader.cc files. Numerous code formatting modifications were performed for clarity and readability. The structure of the class has been modified to enhance its functionality, such as revising how shader logging events are registered, storing the required attributes specifically for Shader, and managing the shader's thread-local object state. These changes were necessary for enhanced thread-safety, and general code clarity. Additionally, a typo in normalMap.h and Map.h  and few other files have been corrected.

* Refactor Shader logging to avoid race conditions

Shader.cc and LoggingAssert.h are added to scene/rdl2/SConscript and CMakeLists.txt to include it in the compilation process. The main change made was in the Shader.h and Shader.cc files. Numerous code formatting modifications were performed for clarity and readability. The structure of the class has been modified to enhance its functionality, such as revising how shader logging events are registered, storing the required attributes specifically for Shader, and managing the shader's thread-local object state. These changes were necessary for enhanced thread-safety, and general code clarity. Additionally, a typo in normalMap.h and Map.h  and few other files have been corrected.

* Rename logging switch for clarity

Changed the variable `mLoggingGlobalSwitch` to `mLoggingEnabled` in the LogEventRegistry class. The method `setLoggingGlobalSwitch( )` has been renamed to `setLoggingEnabled( )` and `getLoggingGlobalSwitch( )` to `getLoggingEnabled( )` respectively.
The modification aims to enhance readability, as `mLoggingEnabled` better communicates the purpose of the variable and associated methods - to enable or disable logging.